+++ changelog.md
@@ -2,7 +2,12 @@
 
 ## Unreleased
 
+### Features
+
+* [1789](https://github.com/zeta-chain/node/issues/1789) - block cross-chain transactions that involve restricted addresses
+
 ### Refactor
+
 * [1783](https://github.com/zeta-chain/node/pull/1783) - refactor zetaclient metrics naming and structure
 
 ### Tests